+++ Vehicles/Car.h
@@ -15,7 +15,7 @@
 class Car : public Vehicle
 {
 public:
-    Car(uint16 number_, path path_, Lane* lane_, DriverType driver_type_);
+    Car(uint32 number_, path path_, Lane* lane_, DriverType driver_type_);
     ~Car();
 private:
 };
